work on create_app structure
