+++ .gitignore
@@ -9,6 +9,7 @@ search-date.json
 .env
 .azure/
 settings.py
+!Api/settings.py
 ._env
 ._env_*
 .pytest_cache/
