+++ Api/Endpoints/base_schema.py
@@ -3,20 +3,15 @@
 
 import datetime
 import re
-
 import marshmallow as MM
 from marshmallow.utils import pprint
-from Endpoints.data_manager import DataManager
-from globals import (
-    db_connection_settings,
-    max_datetime,
-    min_datetime,
-    null_uuid,
-    row_to_dict,
-)
-from Endpoints.references import UUID_Reference
-from Models.gebruikers import Gebruikers_Schema
 import uuid
+
+from Api.Endpoints.data_manager import DataManager
+from Api.settings import max_datetime, min_datetime, null_uuid
+from Api.Utils import row_to_dict
+from Api.Endpoints.references import UUID_Reference
+from Api.Models.gebruikers import Gebruikers_Schema
 
 
 class Short_Base_Schema(MM.Schema):
