+++ Api/Models/ambities.py
@@ -2,15 +2,15 @@
 # Copyright (C) 2018 - 2020 Provincie Zuid-Holland
 
 import marshmallow as MM
-from Endpoints.base_schema import Base_Schema
-from Endpoints.references import (Reverse_ID_Reference, Reverse_UUID_Reference,
-                                  UUID_Linker_Schema)
-from Endpoints.validators import HTML_Validate
-from Models.short_schemas import Short_Beleidskeuze_Schema
-
 from sqlalchemy.orm import relationship
 from sqlalchemy import Column, ForeignKey, Integer, String, Unicode
-from db import CommonMixin, db
+
+from Api.Endpoints.base_schema import Base_Schema
+from Api.Endpoints.references import (Reverse_ID_Reference, Reverse_UUID_Reference,
+                                  UUID_Linker_Schema)
+from Api.Endpoints.validators import HTML_Validate
+from Api.Models.short_schemas import Short_Beleidskeuze_Schema
+from Api.database import CommonMixin, db
 
 
 class Beleidskeuze_Ambities(db.Model):
