+++ Api/Models/short_schemas.py
@@ -1,10 +1,8 @@
-from Endpoints.base_schema import Short_Base_Schema
 import marshmallow as MM
 
+from Api.Endpoints.base_schema import Short_Base_Schema
 
 # For reverse lookups we have to define a custom schema in order to prevent circulair imports
-
-
 class Short_Beleidskeuze_Schema(Short_Base_Schema):
     ID = MM.fields.Integer(required=True, obprops=[])
     UUID = MM.fields.UUID(required=True, obprops=[])
