+++ Tests/resources/images/zuidholland_large.png
